clases empresa completado,falta inversion y usuario

+++ src/clases/Inversion.java
@@ -3,7 +3,7 @@ package clases;
 public class Inversion {
 	private Usuario usuario;
 	private double monto;
-	private double porcentaje;
+	private float porcentaje;
 	private String fecha;
 	private Empresa empresa;
